==> praise/3-wordcount.ipynb <==
# !pip install nltk
# nltk.download('stopwords')
from nltk.corpus import stopwords
import nltk
import pandas as pd

df = pd.read_csv('outputs/praise_data.csv')
df

# Join Words
all_praise = ' '.join(df['Reason for dishing'])
praise_words = all_praise.split(' ')

# Filter Words
filtered_words = [word for word in praise_words if word not in stopwords.words('english')]
filtered_words = [word for word in filtered_words if word != '']
filtered_words = [word for word in filtered_words if word != '-']

# Wordcount
wordcount_df = pd.DataFrame(nltk.FreqDist(filtered_words), index=['count']).T.sort_values('count', ascending=False)

wordcount_df.head(50) / wordcount_df.head(50)['count'].sum()
